Add files via upload
Hey there, 
Welcome to my website. 
This is the lightbox I'm trying to have https://codepen.io/ngatmaitan/pen/GGvxeW
As it works for me now, the image opens but without buttons on the lightbox(left,right,exit).
Please let me know if can see the problem or a solution. 
Thank you!