Success!
The lightbox works! Thank you Lilla.
I am missing the arrowns for next, prev and x  for clossing. I don't mind it , but I can see it might be confusing for others.
I modified  the color from white to dark grey for loading time, and removed the white border.
Very happy for the step forward!